requirements.txt
asgiref==3.7.2
Django==4.2.3
django-admin-interface==0.26.0
django-ckeditor==6.6.1
django-colorfield==0.9.0
django-cors-headers==4.2.0
django-dotenv==1.4.2
django-js-asset==2.1.0
djangorestframework==3.14.0
Pillow==10.0.0
psycopg2-binary==2.9.6
python-slugify==8.0.1
pytz==2023.3
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.7.1
